requirements.txt
joblib
pathlib
selenium
fake-useragent
typing
urllib3
tldextract
requests
Pillow